Release connection after transaction if open-in-view is on
One of the issues with open-in-view is that connections are not released for a
long time, which can lead to connection pool starvation.

Sometimes, controller method even makes no queries for seconds and doesn't really
need to keep the connection for itself all that time. For example that is often
the case, when application makes REST call.

This commit tells Hibernate to release connection after transaction is finished
so that other pieces of code could use it, thus making connection pool starvation
less likely to occur.